Project_Files/threat_detector/requirements.txt
elasticsearch==8.9.0
PyYAML==6.0.2

